+++ docs/changelog.rst
@@ -10,6 +10,7 @@ Development
 - EnumField improvements: now `choices` limits the values of an enum to allow
 - Fix deepcopy of EmbeddedDocument #2202
 - Fix error when using precision=0 with DecimalField #2535
+- Add support for regex and whole word text search query #2568
 
 Changes in 0.23.1
 ===========